fixed logger line nums, dynamic controller finished

+++ www/php/fw/FwDynamicController.php
@@ -103,14 +103,27 @@ class FwDynamicController extends FwController {
         $ps = array(
             'id'    => $id,
             'i'     => $item,
-            'add_users_id_name'  => fw::model('Users')->getFullName($item['add_users_id']),
-            'upd_users_id_name'  => fw::model('Users')->getFullName($item['upd_users_id']),
-            'return_url'        => $this->return_url,
-            'related_id'        => $this->related_id,
+            'return_url'    => $this->return_url,
+            'related_id'    => $this->related_id,
         );
+        if ($this->model->field_add_users_id) $ps['add_users_id_name'] = fw::model('Users')->getFullName($item['add_users_id']);
+        if ($this->model->field_upd_users_id) $ps['upd_users_id_name'] = fw::model('Users')->getFullName($item['upd_users_id']);
+
+        if ($this->is_dynamic_showform) $ps["fields"] = $this->prepareShowFormFields($item, $ps);
+
         if ($this->fw->GLOBAL['ERR']) logger($this->fw->GLOBAL['ERR']);
 
         return $ps;
+    }
+
+    public function modelAddOrUpdate($id, $fields){
+        if ($this->is_dynamic_showform) $this->processSaveShowFormFields($id, $fields);
+
+        $id = parent::modelAddOrUpdate($id, $fields);
+
+        if ($this->is_dynamic_showform) $this->processSaveShowFormFieldsAfter($id, $fields);
+
+        return $id;
     }
 
     public function SaveAction($form_id) {
@@ -139,7 +152,9 @@ class FwDynamicController extends FwController {
     }
 
     public function Validate($id, $item) {
-        $result= $this->validateRequired($item, $this->required_fields);
+        $result= $this->validateRequiredDynamic($item);
+
+        if ($result && $this->is_dynamic_showform) $this->validateSimpleDynamic($id, $item);
 
 /*
         if ($result){
@@ -151,13 +166,69 @@ class FwDynamicController extends FwController {
         $this->validateCheckResult();
     }
 
+    protected function validateRequiredDynamic($item) {
+        $result = true;
+
+        if (!$this->required_fields && $this->is_dynamic_showform) {
+            #if required_fields not defined - fill from showform_fields
+            $fields = $this->config["showform_fields"];
+            $req=array();
+            foreach ($fields as $def) {
+                if ($def['required']) $req[]=$def['field'];
+            }
+            $result = $this->validateRequired($item, $req);
+
+        }else{
+            $result = $this->validateRequired($item, $this->required_fields);
+        }
+
+        return $result;
+    }
+
+    #simple validation via showform_fields
+    public function validateSimpleDynamic($id, $item){
+        $result=true;
+
+        $fields = $this->config["showform_fields"];
+        foreach ($fields as $def) {
+            $field=$def['field'];
+            if (!$field) continue;
+
+            $val = Utils::qh($def["validate"]);
+            if (array_key_exists('exists', $val) && $this->model->isExists( $item[$field], $id ) ){
+                $this->setError($field, 'EXISTS');
+                $result=false;
+            }
+            if (array_key_exists('isemail', $val) && !FormUtils::isEmail($item[$field]) ){
+                $this->setError($field, 'WRONG');
+                $result=false;
+            }
+            if (array_key_exists('isphone', $val) && !FormUtils::isPhone($item[$field]) ){
+                $this->setError($field, 'WRONG');
+                $result=false;
+            }
+            if (array_key_exists('isdate', $val) && !FormUtils::isDate($item[$field]) ){
+                $this->setError($field, 'WRONG');
+                $result=false;
+            }
+            if (array_key_exists('isfloat', $val) && !FormUtils::isFloat($item[$field]) ){
+                $this->setError($field, 'WRONG');
+                $result=false;
+            }
+
+            #if (!$result) break; #uncomment to break on first error
+        }
+
+        return $result;
+    }
+
     public function ShowDeleteAction($id){
         $id+=0;
         $ps = array(
             'i' => $this->model->one($id),
             'return_url'        => $this->return_url,
             'related_id'        => $this->related_id,
-            'base_url'          => $this->fw->config->ROOT_URL.$this->base_url, #override default template url, remove if you created custom /showdelete templates
+            'base_url'          => $this->base_url, #override default template url, remove if you created custom /showdelete templates
         );
 
         $this->fw->parser('/common/form/showdelete', $ps);
@@ -176,16 +247,31 @@ class FwDynamicController extends FwController {
         $acb = req('cb');
         if (!is_array($acb)) $acb=array();
         $is_delete = reqs('delete')>'';
+        $user_lists_id  = reqi("addtolist");
+        $remove_user_lists_id = reqi("removefromlist");
+
+        if ($user_lists_id) {
+            $user_lists = fw::model('UserLists')->one($user_lists_id);
+            if (!$user_lists || $user_lists["add_users_id"] <> Utils::me()) throw New ApplicationException("Wrong Request");
+        }
 
         $ctr=0;
         foreach ($acb as $id => $value) {
             if ($is_delete){
                 $this->model->delete($id);
                 $ctr+=1;
+            }elseif ($user_lists_id){
+                fw::model('UserLists').addItemList($user_lists_id, $id);
+                $ctr += 1;
+            }elseif ($remove_user_lists_id){
+                fw::model('UserLists').delItemList($remove_user_lists_id, $id);
+                $ctr += 1;
             }
         }
 
         $this->fw->flash("multidelete", $ctr);
+        if ($user_lists_id) fw.FLASH("success", "$ctr records added to the list");
+
         fw::redirect($this->getReturnLocation());
     }
 
@@ -195,6 +281,47 @@ class FwDynamicController extends FwController {
         $items = $this->model_related->getAutocompleteList(reqs("q"));
 
         return array('_json' => $items);
+    }
+
+    ###################### support for customizable list screen
+    public function UserViewsAction($id){
+        $id+=0;
+
+        $ps=array(
+            'rows' => $this->getViewListArr($this->getViewListUserFields(), true)
+        );
+
+        $this->fw->parser("/common/list/userviews", $ps);
+    }
+
+    public function SaveUserViewsAction(){
+        $item = reqh('item');
+        $success = true;
+
+        try {
+            if (reqi("is_reset")) {
+                fw::model('UserViews')->updateByScreen($this->base_url, $this->view_list_defaults);
+            }else{
+                #save fields
+                #order by value
+                $ordered = reqh("fld");
+                asort($ordered);
+
+                #and then get ordered keys
+                $anames = array();
+                foreach ($ordered as $key => $value) {
+                    $anames[]=$key;
+                }
+
+                fw::model('UserViews')->updateByScreen($this->base_url, implode(' ', $anames));
+            }
+
+        } catch (Exception $ex) {
+            $success=false;
+            $this->setFormError($ex->getMessage());
+        }
+
+        fw::redirect($this->return_url);
     }
 
     ###################### HELPERS for dynamic fields
@@ -209,6 +336,7 @@ class FwDynamicController extends FwController {
         $id = $item['id']+0;
 
         $fields = $this->config["show_fields"];
+        if (!$fields) throw New ApplicationException("Controller config.json doesn't contain 'show_fields'");
         foreach ($fields as &$def) {
             $def['i'] = $item;
             $dtype = $def["type"];
@@ -227,7 +355,6 @@ class FwDynamicController extends FwController {
 
             }elseif ($dtype == "att_links"){
                 $def["att_links"] = fw::model('Att')->getAllLinked($this->model->table_name, $id);
-                logger($def["att_links"]);
 
             }else{
                 #single values
@@ -270,6 +397,144 @@ class FwDynamicController extends FwController {
         unset($def);
 
         return $fields;
+    }
+
+    /**
+     * prepare data for fields repeat in ShowFormAction based on config.json showform_fields parameter
+     * @param  array $item one item
+     * @param  array $ps   for parsepage
+     * @return array       array of hashtables to build fields in templates
+     */
+    public function prepareShowFormFields($item, $ps){
+        $id = $item['id']+0;
+
+        $fields = $this->config["showform_fields"];
+        if (!$fields) throw New ApplicationException("Controller config.json doesn't contain 'showform_fields'");
+        foreach ($fields as &$def) {
+            $def['i'] = $item;  #ref to item
+            $def['ps'] = $ps;   #ref to whole ps
+            $dtype = $def["type"]; #type is required
+            $field = $def["field"];
+
+            if ($dtype == "row" || $dtype == "row_end" || $dtype == "col" || $dtype == "col_end"){
+                #structural tags
+                $def["is_structure"] = true;
+
+            }elseif ($dtype == "multicb"){
+                #complex field
+                $def["multi_datarow"] = fw::model($def["lookup_model"])->getMultiList($item[$field], $def["lookup_params"]);
+                foreach ($def["multi_datarow"] as &$row) {
+                    $row["field"] = $def["field"];
+                }
+                unset($row);
+
+            }elseif ($dtype == "att_edit"){
+                $def["att"] = fw::model('Att')->one($item[$field]);
+                $def["value"] = $item[$field];
+
+            }elseif ($dtype == "att_links_edit"){
+                $def["att_links"] = fw::model('Att')->getAllLinked($this->model->table_name, $id);
+
+            }else{
+                #single values
+                #lookups
+                if (array_key_exists('lookup_table', $def)){
+                    #lookup by table
+                    $lookup_key = $def["lookup_key"];
+                    if (!$lookup_key) $lookup_key = "id";
+
+                    $lookup_field = $def["lookup_field"];
+                    if (!$lookup_field) $lookup_field = "iname";
+
+                    $def["lookup_row"] = $this->db->row($def["lookup_table"], array($lookup_key => $item[$field]) );
+                    $def["value"] = $def["lookup_row"][$lookup_field];
+
+                }elseif(array_key_exists('lookup_model', $def)){
+                    if (array_key_exists('lookup_field', $def)){
+                        #lookup value
+                        $def["lookup_row"] = fw::model($def["lookup_model"])->one($item[$field]);
+                        $def["value"] = $def["lookup_row"][$lookup_field];
+                    }else{
+                        #lookup select
+                        $def["select_options"] = fw::model($def["lookup_model"])->listSelectOptions($def['lookup_params']);
+                        $def["value"] = $item[$field];
+                    }
+
+                }elseif(array_key_exists('lookup_tpl', $def)){
+                    $def['select_options'] = FormUtils::selectTplOptions($def['lookup_tpl'], $item[$field]);
+                    $def["value"] = $item[$field];
+                    foreach ($def['select_options'] as &$row) { #contains id, iname
+                        $row["is_inline"] = $def["is_inline"];
+                        $row["field"] = $def["field"];
+                        $row["value"] = $item["field"];
+                    }
+                    unset($row);
+                }else{
+                    $def["value"] = $item[$field];
+                }
+            }
+
+            #convertors
+            if (array_key_exists('conv', $def)){
+                if ($def["conv"] == "time_from_seconds"){
+                    $def["value"] = DateUtils::int2timestr($def["value"]);
+                }
+            }
+        }
+        unset($def);
+
+        return $fields;
+    }
+
+    #auto-process fields BEFORE record saved to db
+    public function processSaveShowFormFields($id, &$fields){
+        $item = reqh("item");
+
+        $showform_fields = $this->_fieldsToHash($this->config["showform_fields"]);
+        #special auto-processing for fields of particular types
+        foreach (array_keys($fields) as $field) {
+            if (array_key_exists($field, $showform_fields)){
+                $def = $showform_fields[$field];
+                if ($def['type']=='multicb'){
+                    #multiple checkboxes
+                    $fields[$field] = FormUtils::multi2ids(reqh($field."_multi"));
+                }elseif ($def['type']=='autocomplete'){
+                    $fields[$field] = fw::model($def["lookup_model"])->findOrAddByIname($fields[$field]);
+                }elseif ($def['type']=='date_combo'){
+                    $fields[$field] = FormUtils::dateForCombo($item, $field);
+                }elseif ($def['type']=='date_popup'){
+                    $fields[$field] = DateUtils::Str2SQL($fields[$field]); #convert date to sql format
+                }elseif ($def['type']=='time'){
+                    $fields[$field] = DateUtils::timestr2int($fields[$field]); #ftime - convert from HH:MM to int (0-24h in seconds)
+                }elseif ($def['type']=='number'){
+                    $fields[$field] = floatval($fields[$field]); #number - convert to number (if field empty or non-number - it will become 0)
+                }
+            }
+        }
+
+        #$fields["fint"]=$fields["fint"]+0 'TODO? field accepts only int
+    }
+
+    #auto-process fields AFTER record saved to db
+    protected function processSaveShowFormFieldsAfter($id, $fields){
+        #for now we just look if we have att_links_edit field and update att links
+        foreach ($this->config['showform_fields'] as $def) {
+            if ($def['type']=='att_links_edit'){
+                fw::model('Att')->updateAttLinks($this->model->table_name, $id, reqh("att")); #TODO make att configurable
+            }
+        }
+    }
+
+    #convert config's fields list into hashtable as field => {}
+    #if there are more than one field - just first field added to the hash
+    protected function _fieldsToHash($fields){
+        $result=array();
+        foreach ($fields as $fldinfo) {
+            if (array_key_exists('field', $fldinfo) && !array_key_exists($fldinfo['field'], $result)){
+                $result[$fldinfo['field']]=$fldinfo;
+            }
+        }
+        return $result;
     }
 
 }//end of class
